Change menu icons and split files.
