+++ frontend/src/app/chat/index.js
@@ -4,4 +4,5 @@ export * from './ChatConstants'
 export * from './ChatConversation'
 export * from './ChatHeader'
 export * from './ChatHooks'
+export * from './ChatMenu'
 export * from './Decorators'
