+++ frontend/src/pages/ChatPage.jsx
@@ -4,6 +4,14 @@ import { useParams } from 'react-router-dom'
 import { ChatApp, ChatHeader, ChatConversation, ChatComposer } from '../app/chat'
 import { COMMUNITY } from '../app/constants'
 
+const MENU_CONFIG = {
+    options: [
+        { id: 'search', name: 'Search Messages' },
+        { id: 'block', name: 'Block Users' },
+        { id: 'rate', name: 'Rate Match' },
+    ],
+}
+
 export default function ChatPage() {
     const { chatId } = useParams()
     const fullChatId = `${COMMUNITY}/${chatId}`
@@ -12,7 +20,7 @@ export default function ChatPage() {
             <div className="ChatContainer">
                 <ChatApp
                     chatId={fullChatId}
-                    header={<ChatHeader />}
+                    header={<ChatHeader menuConfig={MENU_CONFIG} />}
                     conversation={<ChatConversation />}
                     composer={<ChatComposer />}
                 />
